# backend/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Verified B2B Deal Network API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistrationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Auth token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── Organizations ───────────────────────────────────────────────────────────
...
  /admin/verification-cases:
    get:
      operationId: adminListVerificationCases
      tags: [admin]
      summary: List all verification cases (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VerificationCase"
  /admin/verification-cases/{caseId}:
    get:
      operationId: adminGetVerificationCase
      tags: [admin]
      summary: Get a verification case with documents (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Case with documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationCaseDetail"
  /admin/verification-cases/{caseId}/decide:
    post:
      operationId: adminDecideVerificationCase
      tags: [admin]
      summary: Approve, reject, or request more info on a verification case
      security:
        - bearerAuth: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationDecisionInput"
      responses:
        "200":
          description: Updated case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationCase"

  # ─── Connections ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: false
      properties:
        error:
          type: string
      required:
        - error

    # ── Auth ────────────────────────────────────────────────────────────────────
    UserRegistrationInput:
      type: object
      additionalProperties: false
      required: [email, password, name]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 8
        name:
          type: string
    LoginInput:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      additionalProperties: false
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      additionalProperties: false
      required: [id, email, name, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time

    # ── Organizations ────────────────────────────────────────────────────────────
...
    VerificationCase:
      type: object
      additionalProperties: false
      required: [id, organizationId, status, createdAt]
      properties:
        id:
          type: string
        organizationId:
          type: string
        status:
          type: string
          enum: [SUBMITTED, IN_REVIEW, APPROVED, REJECTED, MORE_INFO_NEEDED]
        reviewerAdminId:
          type: ["string", "null"]
        decisionNotes:
          type: ["string", "null"]
        decidedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    VerificationCaseDetail:
      allOf:
        - $ref: "#/components/schemas/VerificationCase"
        - type: object
          additionalProperties: false
          properties:
            documents:
              type: array
              items:
                $ref: "#/components/schemas/DocumentRef"
    VerificationDecisionInput:
      type: object
      additionalProperties: false
      required: [decision]
      properties:
        decision:
          type: string
          enum: [APPROVED, REJECTED, MORE_INFO_NEEDED]
        notes:
          type: string
...
    DocumentRef:
      type: object
      additionalProperties: false
      required: [id, organizationId, docType, ipfsCid, contentHash, uploadedBy, uploadedAt]
      properties:
        id:
          type: string
        organizationId:
          type: string
        docType:
          type: string
          enum: [REGISTRATION_CERT, ID, PROOF_OF_ADDRESS, AGREEMENT, INVOICE, EVIDENCE, OTHER]
        ipfsCid:
          type: string
        contentHash:
          type: string
        uploadedBy:
          type: string
        uploadedAt:
          type: string
          format: date-time
...